feat(backends): float32 precision knob and leaner jax focal-plane builds
New top-level 'precision' config field (jax backend only; default
float64): builds the MFT propagation kernels in single precision for
half-memory, faster sampling. Images and reference PSFs come out
float32 — the intensity output tap now preserves the propagation dtype
instead of forcing float64. Requesting float32 with the hcipy backend
is a config error.

The jax focal planes no longer construct the unused hcipy propagator
and wavefront objects at build time (their lam_setup.propagator is now
None and per-wavefront propagate() raises clearly), cutting build
overhead; complex aperture fields are rejected with a clear error since
the MFT path propagates a real transmission amplitude.

+++ src/telescope_sim/outputs/intensity.py
@@ -61,9 +61,9 @@ class IntensityOutputTap(OutputTap):
         missing = [n for n in self.focal_plane_names if n not in fp_results]
         if missing:
             raise KeyError(f"focal planes {missing} not in available outputs {list(fp_results)}")
-        psfs = [
-            np.asarray(fp_results[n].intensity, dtype=np.float64) for n in self.focal_plane_names
-        ]
+        # Preserve the propagation dtype: float64 on the default path, or
+        # float32 when the jax backend runs with precision="float32".
+        psfs = [np.asarray(fp_results[n].intensity) for n in self.focal_plane_names]
         return np.stack(psfs, axis=-1)
